Lock std::cout + debug method for missed sequences
If a sequences was missed, so diff(called_sequence, sequence) > 1), the
caller must check for that and adjust the counters accordingly.
This simple example is not doing that, so when DELAY_TEST is set to 1,
the application will hang since thread[1] is still waiting for one
round to be completed.
Since the producer is not producing anymore it will hang in
wait_until_full forever.

Signed-off-by: Maximilian Güntner <maximilian.guentner@gmail.com>